retag_requirements.txt
git+https://github.com/cedadev/ceda-elasticsearch-tools.git
certifi==2020.4.5.2
chardet==3.0.4
docopt==0.6.2
elasticsearch==7.6.0
gitdb==4.0.5
GitPython==3.1.3
idna==2.9
pika==1.1.0
requests==2.23.0
smmap==3.0.4
urllib3==1.25.9
